+++ res/res_monitor.c
@@ -150,8 +150,8 @@ int ast_monitor_start(	struct ast_channel *chan, const char *format_spec,
 				while((p = strchr(channel_name, '/'))) {
 					*p = '-';
 				}
-				snprintf(monitor->filename_base, FILENAME_MAX, "%s/%ld-%s",
-						 ast_config_AST_MONITOR_DIR, time(NULL),channel_name);
+				snprintf(monitor->filename_base, FILENAME_MAX, "%s/%d-%s",
+						 ast_config_AST_MONITOR_DIR, (int)time(NULL),channel_name);
 				monitor->filename_changed = 1;
 			} else {
 				ast_log(LOG_ERROR,"Failed to allocate Memory\n");
